feat: API URL path 수정 및 MemberController 역할 분리
